# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API for creating Questions and Options",
    "title": "Leaderboard"
  },
  "servers": [
    {
      "url": "https://blacklight-assg.onrender.com"
    }
  ],
  "paths": {
    "/leaderboard/current-week": {
      "get": {
        "tags": ["Current-Week"],
        "summary": "Get current-Week leaderboard",
        "description": "Retrieve the current week's leaderboard",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "UID": { "type": "string" },
                      "Name": { "type": "string" },
                      "Score": { "type": "integer" },
                      "Country": { "type": "string" },
                      "TimeStamp": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/last-week/{country}": {
      "get": {
        "tags": ["Last-Week"],
        "summary": "Get last week leaderboard by country",
        "description": "Retrieve last week's leaderboard for a specific country",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "required": true,
            "description": "ISO 2 letter country code",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "UID": { "type": "string" },
                      "Name": { "type": "string" },
                      "Score": { "type": "integer" },
                      "Country": { "type": "string" },
                      "TimeStamp": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/rank": {
      "get": {
        "tags": ["User-Rank"],
        "summary": "Get user rank by userId",
        "description": "Retrieve the rank of a user based on their userId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rank": { "type": "integer", "description": "User rank" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}